test/functional/pyproject.toml
[tool.poetry]
name = "updatescanner_functional_tests"
version = "0.0.0"
description = "Update Scanner Functional Tests"
authors = ["sneakypete81 <sneakypete81@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.7"
flake8 = "^3.7"
pytest = "^4.6"
pyobjc = "^5.2" # TODO: Remove!
PyHamcrest = "^1.9"
requests = "^2.22"

clickshot = "0.1.0"
opencv-python = "^4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
